refactor(types-first-dev): Flatten module tree, remove V3 prefix
The V3 naming is now meaningless since it's the only implementation.
Flatten the module hierarchy for simpler imports.

Changes:
- Move TypesFirstDev.V3.* → TypesFirstDev.*
- Move templates/v3/* → templates/*
- Rename V3Graph → TDDGraph, V3Assessment → TDDAssessment
- Update all imports and module declarations
- Clean up V3 references in comments

Final structure:
  src/TypesFirstDev/
    Types.hs, Types/{Core,Shared,Payloads,Memory,Nodes}.hs
    Context.hs, Templates.hs, Graph.hs
    Policy.hs, Evaluation.hs, Decomposition.hs
    Handlers.hs, Schema.hs
