+++ types-first-dev/src/TypesFirstDev/Handlers.hs
@@ -1,9 +1,9 @@
 -- | Handlers for the types-first development workflow.
 --
--- V3 handlers are not yet implemented. This module provides stubs.
--- See V3.Policy for routing decision logic.
+-- Handlers are not yet implemented. This module provides stubs.
+-- See Policy for routing decision logic.
 module TypesFirstDev.Handlers
-  ( -- * V3 Handler Stubs
+  ( -- * Handler Stubs
     maxRetryAttempts
   ) where
