+++ types-first-dev/src/TypesFirstDev/Types.hs
@@ -1,21 +1,27 @@
--- | Schema types for the types-first development workflow.
+-- | TDD Protocol Types
 --
--- V3 is the canonical implementation. Re-exports from V3.Types.
+-- Re-exports all types from submodules.
+-- Field prefix convention: lowercase acronym of type name (e.g., Spec -> s)
+--
+-- Design principles:
+-- 1. LLM outputs describe WHAT was done, not the code itself
+-- 2. Types derive StructuredOutput for unified schema+parsing
+-- 3. Memory types enable threaded conversation context
 module TypesFirstDev.Types
-  ( -- * Core Types
-    module TypesFirstDev.V3.Types.Core
-    -- * Shared Types
-  , module TypesFirstDev.V3.Types.Shared
-    -- * Payload Types
-  , module TypesFirstDev.V3.Types.Payloads
-    -- * Memory Types
-  , module TypesFirstDev.V3.Types.Memory
-    -- * Node Types
-  , module TypesFirstDev.V3.Types.Nodes
+  ( -- * Core Entry Types
+    module TypesFirstDev.Types.Core
+    -- * Shared Workflow Types
+  , module TypesFirstDev.Types.Shared
+    -- * Cross-Node Payloads
+  , module TypesFirstDev.Types.Payloads
+    -- * Node-Private Memory
+  , module TypesFirstDev.Types.Memory
+    -- * Node Input/Output Types
+  , module TypesFirstDev.Types.Nodes
   ) where
 
-import TypesFirstDev.V3.Types.Core
-import TypesFirstDev.V3.Types.Shared
-import TypesFirstDev.V3.Types.Payloads
-import TypesFirstDev.V3.Types.Memory
-import TypesFirstDev.V3.Types.Nodes
+import TypesFirstDev.Types.Core
+import TypesFirstDev.Types.Shared
+import TypesFirstDev.Types.Payloads
+import TypesFirstDev.Types.Memory
+import TypesFirstDev.Types.Nodes
