+++ types-first-dev/templates/SPEC.md
