+++ types-first-dev/templates/scaffold.jinja
