+++ types-first-dev/test/Main.hs
@@ -1,12 +1,12 @@
--- | V3 test suite (placeholder).
+-- | Test suite (placeholder).
 --
--- Tests will be added as V3 handlers are implemented.
+-- Tests will be added as handlers are implemented.
 module Main (main) where
 
 import Test.Hspec
 
 main :: IO ()
 main = hspec $ do
-  describe "V3 Types" $ do
+  describe "TypesFirstDev" $ do
     it "placeholder test" $ do
       True `shouldBe` True
